requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0,<3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
langchain_google_genai==2.1.4
langchain
black